Split codex.py into _review, _doctor, and _status modules
scripts/codex.py drops from 2437 to 620 lines: review
orchestration moves to _review.py (1088), doctor/preflight checks
to _doctor.py (641), status rendering to _status.py (305).
codex.py keeps CLI parsing, routing, shared config helpers, and
re-exports every previously-public name so 'import codex' callers
and test monkeypatching keep working — constants that tests patch
on the codex module (_LIVE_PROBE_TIMEOUT, _REVIEW_PASS_THRESHOLD,
_rm) are read back through the codex namespace lazily at call
time. tests/test_codex_module_split.sh gains 19 assertions for
the new layout.

Closes #206
