+++ scripts/codex.py
@@ -2,25 +2,18 @@
 """Codex-native Trinity command wrapper."""
 
 import argparse
-import datetime as dt
-import difflib
-import hashlib
 import json
 import os
 from pathlib import Path
-import re
-import shlex
-import shutil
-import signal
 import subprocess
 import sys
 
 try:
     from ._version import load_version
-    from . import _review_metadata as _rm
+    from . import _review_metadata as _rm  # noqa: F401 — tests patch codex._rm
 except ImportError:
     from _version import load_version
-    import _review_metadata as _rm
+    import _review_metadata as _rm  # noqa: F401 — tests patch codex._rm
 
 try:
     from . import provider_runtime as _provider_runtime
@@ -31,6 +24,21 @@ try:
     from . import review_schema as _review_schema
 except ImportError:
     import review_schema as _review_schema
+
+try:
+    from . import _doctor as _doctor_mod
+except ImportError:
+    import _doctor as _doctor_mod
+
+try:
+    from . import _review as _review_mod
+except ImportError:
+    import _review as _review_mod
+
+try:
+    from . import _status as _status_mod
+except ImportError:
+    import _status as _status_mod
 
 parse_provider_command = _provider_runtime.parse_provider_command
 provider_timeout = _provider_runtime.provider_timeout
@@ -58,6 +66,84 @@ _strip_stderr_region = _review_schema._strip_stderr_region
 _safe_read_raw = _review_schema._safe_read_raw
 _validate_review_schema = _review_schema._validate_review_schema
 _REVIEW_PASS_THRESHOLD = _review_schema._REVIEW_PASS_THRESHOLD
+
+
+# Re-exports from _doctor module (doctor/preflight/health-check subcommand).
+_MIN_TIMEOUT_WARNING_SECONDS = _doctor_mod._MIN_TIMEOUT_WARNING_SECONDS
+_WRAPPER_AUTH_CONFIG = _doctor_mod._WRAPPER_AUTH_CONFIG
+_make_health_result = _doctor_mod._make_health_result
+wrapper_auth_check = _doctor_mod.wrapper_auth_check
+_is_canonical_wrapper = _doctor_mod._is_canonical_wrapper
+detect_env_pollution = _doctor_mod.detect_env_pollution
+executable_health = _doctor_mod.executable_health
+provider_health = _doctor_mod.provider_health
+provider_health_results = _doctor_mod.provider_health_results
+_format_provider_block = _doctor_mod._format_provider_block
+format_health_results = _doctor_mod.format_health_results
+health_results_ok = _doctor_mod.health_results_ok
+# NOTE: _LIVE_PROBE_TIMEOUT stays as a real attribute on this module (NOT
+# re-exported from _doctor) so monkeypatch.setattr(codex_mod,
+# "_LIVE_PROBE_TIMEOUT", ...) in test_doctor_preflight.py takes effect.
+# _doctor._probe_provider reads this value via lazy `import codex` at call time.
+_LIVE_PROBE_TIMEOUT = 10  # hard 10s timeout for live probes.
+_probe_provider = _doctor_mod._probe_provider
+cmd_doctor = _doctor_mod.cmd_doctor
+
+# Re-exports from _review module (review orchestration subcommand).
+scope_pathspec = _review_mod.scope_pathspec
+git_diff = _review_mod.git_diff
+git_diff_range = _review_mod.git_diff_range
+changed_paths = _review_mod.changed_paths
+changed_paths_range = _review_mod.changed_paths_range
+untracked_paths = _review_mod.untracked_paths
+is_text_file = _review_mod.is_text_file
+read_text_file = _review_mod.read_text_file
+read_text_file_at_ref = _review_mod.read_text_file_at_ref
+synthetic_untracked_diff = _review_mod.synthetic_untracked_diff
+file_snapshots = _review_mod.file_snapshots
+file_snapshots_at_ref = _review_mod.file_snapshots_at_ref
+review_input_sha256 = _review_mod.review_input_sha256
+make_review_input = _review_mod.make_review_input
+working_tree_review_input = _review_mod.working_tree_review_input
+base_head_review_input = _review_mod.base_head_review_input
+run_gh = _review_mod.run_gh
+gh_pr_view = _review_mod.gh_pr_view
+gh_pr_diff = _review_mod.gh_pr_diff
+git_commit_exists = _review_mod.git_commit_exists
+ensure_pr_head_available = _review_mod.ensure_pr_head_available
+pr_changed_paths = _review_mod.pr_changed_paths
+pr_review_input = _review_mod.pr_review_input
+resolve_review_input = _review_mod.resolve_review_input
+normalize_review_doc_id = _review_mod.normalize_review_doc_id
+strict_review_metadata = _review_mod.strict_review_metadata
+resolve_strict_review = _review_mod.resolve_strict_review
+render_strict_review_instructions = _review_mod.render_strict_review_instructions
+render_prompt = _review_mod.render_prompt
+slugify = _review_mod.slugify
+make_review_dir = _review_mod.make_review_dir
+review_parallelism = _review_mod.review_parallelism
+ordered_subset = _review_mod.ordered_subset
+ordered_cleanup = _review_mod.ordered_cleanup
+write_incomplete = _review_mod.write_incomplete
+_sanitize_md = _review_mod._sanitize_md
+_render_finding_items = _review_mod._render_finding_items
+_render_findings_for = _review_mod._render_findings_for
+_compute_summary = _review_mod._compute_summary
+_render_summary_block = _review_mod._render_summary_block
+_format_cli_summary = _review_mod._format_cli_summary
+write_synthesis = _review_mod.write_synthesis
+STRICT_REVIEW_OUTPUT_SCHEMA = _review_mod.STRICT_REVIEW_OUTPUT_SCHEMA
+REVIEW_ONLY_INSTRUCTION = _review_mod.REVIEW_ONLY_INSTRUCTION
+STRICT_REVIEW_TEMPLATES = _review_mod.STRICT_REVIEW_TEMPLATES
+cmd_review = _review_mod.cmd_review
+
+# Re-exports from _status module (status subcommand).
+_format_elapsed = _status_mod._format_elapsed
+_format_duration = _status_mod._format_duration
+_format_ago = _status_mod._format_ago
+_print_incomplete_only_summary = _status_mod._print_incomplete_only_summary
+_print_review_summary = _status_mod._print_review_summary
+cmd_status = _status_mod.cmd_status
 
 
 def parse_structured_review(raw_text, pass_threshold=None):
@@ -118,45 +204,6 @@ PRESET_ALIAS_RESERVED_WORDS = {
     "help",
     "session-path",
 }
-STRICT_REVIEW_OUTPUT_SCHEMA = [
-    "### Findings",
-    "### Decision Matrix",
-    "**Weighted Average: X.X/10 - PASS/FIX**",
-]
-REVIEW_ONLY_INSTRUCTION = (
-    "## Review-Only Mode\n\n"
-    "Do not run tests, shell commands, network calls, or mutate files unless the "
-    "instructions in this review prompt explicitly ask you to. Base findings only "
-    "on the provided diff, file snapshots, and review context.\n"
-)
-STRICT_REVIEW_TEMPLATES = {
-    ("COR-1602", "COR-1609"): {
-        "pass_threshold": 9.0,
-        "decision_rule": (
-            "PASS when weighted_average >= 9.0 and no blocking findings remain; "
-            "otherwise FIX."
-        ),
-        "calibration": "COR-1611",
-        "rubric_title": "CHG Review Scoring",
-        "criteria": [
-            ("Correctness", "25%", "Change is technically sound and feasible."),
-            ("Completeness", "25%", "CHG covers scope, risks, verification, and docs."),
-            (
-                "TDD Plan Quality",
-                "20%",
-                "Plan uses RED/GREEN/REFACTOR where code changes are involved.",
-            ),
-            ("Consistency", "15%", "Fits local conventions and related COR/TRN docs."),
-            ("Rollback Safety", "15%", "Rollback path and blast radius are clear."),
-        ],
-        "non_code_note": (
-            "For non-code CHGs where TDD is not applicable, redistribute the TDD "
-            "weight to Completeness (35%) and Consistency (25%)."
-        ),
-    }
-}
-
-
 __version__ = load_version()
 
 
@@ -422,1870 +469,6 @@ def resolve_review_providers(args, config):
         },
         [],
     )
-
-
-_MIN_TIMEOUT_WARNING_SECONDS = 60
-
-# TRN-3021: per-provider auth-file paths for wrapper providers. Hardcoded
-# v1 because exactly 2 providers in providers/registry.json use this
-# pattern (verified: `grep KEY_FILE providers/bin/*` returns deepseek + openrouter).
-# Tagged tech-debt — declarative `auth_env` + `auth_file` registry fields
-# belong in TRN-3030 follow-up CHG.
-_WRAPPER_AUTH_CONFIG = {
-    "deepseek": {
-        "env_var": "DEEPSEEK_API_KEY",
-        "file": "~/.secrets/deepseek_api_key",
-    },
-    "openrouter": {
-        "env_var": "OPENROUTER_API_KEY",
-        "file": "~/.secrets/openrouter_api_key",
-    },
-}
-
-
-def _make_health_result(
-    provider,
-    *,
-    ok,
-    issues=(),
-    warnings=(),
-    executable=None,
-    timeout=None,
-    cli=None,
-    auth=None,
-    timeout_warning=False,
-):
-    """Construct a provider_health result dict with all TRN-3021 keys.
-
-    Used by all 3 sites that build result dicts: canonical at
-    provider_health (line ~413), synthetic-error branches at
-    provider_health_results (lines ~425, ~440). Defensive defaults
-    keep the dict additive — older consumers reading via .get() see
-    sensible values for new keys.
-    """
-    return {
-        "provider": provider,
-        "ok": ok,
-        "executable": executable,
-        "timeout": timeout,
-        "issues": list(issues),
-        # TRN-3021 additions:
-        "warnings": list(warnings),
-        "cli": cli,
-        "auth": auth,
-        "timeout_warning": timeout_warning,
-        "live_probe": None,
-    }
-
-
-def wrapper_auth_check(provider):
-    """TRN-3021: check wrapper-provider auth via env-or-file precedence.
-
-    Mirrors providers/bin/<provider>:10-31 contract:
-      1. env var (DEEPSEEK_API_KEY / OPENROUTER_API_KEY) → source: "env"
-      2. file at ~/.secrets/<provider>_api_key with mode 600 or 400
-         → source: "file", mode_ok: True
-      3. file with wrong mode → source: "file", mode_ok: False
-         (wrapper exits 1 at runtime; severity is fatal for REQUIRED)
-      4. neither → source: "missing"
-
-    Returns None for non-wrapper providers (vendor login flow, no file).
-    """
-    if provider not in _WRAPPER_AUTH_CONFIG:
-        return None
-    config = _WRAPPER_AUTH_CONFIG[provider]
-    env_name = config["env_var"]
-    if os.environ.get(env_name):
-        return {
-            "source": "env",
-            "env_var": env_name,
-            "file": None,
-            "mode_ok": None,
-        }
-    file_path = Path(config["file"]).expanduser()
-    if not file_path.exists():
-        return {
-            "source": "missing",
-            "env_var": env_name,
-            "file": str(file_path),
-            "mode_ok": None,
-        }
-    try:
-        # Use lstat to match wrapper's `stat -c '%a' "$KEY_FILE"` behavior:
-        # on GNU/Linux that reports the symlink's own mode (typically 777),
-        # not the target's. Following the symlink (file_path.stat()) would
-        # report a 0600 target as ok while the wrapper still refuses to
-        # read the symlink at runtime — inconsistency caught by codex bot
-        # on PR #68.
-        mode = file_path.lstat().st_mode & 0o777
-    except OSError:
-        return {
-            "source": "file",
-            "env_var": env_name,
-            "file": str(file_path),
-            "mode_ok": False,
-        }
-    return {
-        "source": "file",
-        "env_var": env_name,
-        "file": str(file_path),
-        "mode_ok": mode in (0o600, 0o400),
-    }
-
-
-def _is_canonical_wrapper(executable, provider):
-    """True iff `executable` is the actual canonical wrapper script for `provider`.
-
-    The canonical wrapper lives at one of these EXACT paths (exact basename
-    match, no substring fuzzing):
-      - <HOME>/.claude/skills/trinity/bin/<provider>
-      - <HOME>/.codex/skills/trinity/bin/<provider>
-      - <repo>/providers/bin/<provider>  (when running tests from repo root)
-
-    Custom configs that reuse the provider name with a different executable
-    (e.g. `deepseek-test`) or a different path (e.g. `/tmp/skills/trinity/
-    bin/deepseek`) won't match, so cmd_doctor's wrapper-auth check stays
-    out of their way.
-
-    Compares against `Path(executable).resolve()` so a symlinked install
-    still matches the canonical target.
-    """
-    if executable is None or provider not in _WRAPPER_AUTH_CONFIG:
-        return False
-    try:
-        actual = Path(executable).resolve()
-    except (OSError, RuntimeError):
-        return False
-
-    home = Path(os.path.expanduser("~"))
-    candidates = [
-        home / ".claude" / "skills" / "trinity" / "bin" / provider,
-        home / ".codex" / "skills" / "trinity" / "bin" / provider,
-        Path(__file__).resolve().parent.parent / "providers" / "bin" / provider,
-    ]
-
-    for c in candidates:
-        try:
-            if c.resolve() == actual:
-                return True
-        except (OSError, RuntimeError):
-            continue
-    return False
-
-
-def detect_env_pollution(base_env=None):
-    """TRN-3021: scan parent env for vars that match the TRN-3023 clearlist.
-
-    Flag iff `_matches_any(key, _DEFAULT_ENV_CLEAR_PATTERNS) AND NOT
-    _is_essential(key)`. Essentials check first (defensive ordering, matches
-    `build_provider_env` at codex.py:_is_essential). Values truncated to
-    first 12 chars + "…" to avoid leaking corp hostnames / token-adjacent
-    strings. Returns list of (key, redacted_value) sorted alphabetically
-    by key for deterministic display.
-    """
-    if base_env is None:
-        base_env = os.environ
-    found = []
-    for key, value in base_env.items():
-        if _is_essential(key):
-            continue
-        if _matches_any(key, _DEFAULT_ENV_CLEAR_PATTERNS):
-            redacted = value if len(value) <= 12 else value[:12] + "…"
-            found.append((key, redacted))
-    found.sort(key=lambda kv: kv[0])
-    return found
-
-
-def executable_health(command, root):
-    executable = command[0]
-    if command_has_path(executable):
-        executable_path = Path(executable)
-        if not executable_path.is_absolute():
-            executable_path = root / executable_path
-        if not executable_path.exists():
-            return None, f"command not found: {executable_path}"
-        if executable_path.is_dir() or not os.access(executable_path, os.X_OK):
-            return str(executable_path), f"not executable: {executable_path}"
-        return str(executable_path), None
-
-    resolved = shutil.which(executable)
-    if resolved is None:
-        return None, f"command not found: {executable}"
-    return resolved, None
-
-
-def provider_health(provider, provider_config, root):
-    """TRN-3021: extended preflight health check.
-
-    Returns a result dict with keys: provider, ok, executable, timeout,
-    issues, warnings, cli, auth, timeout_warning. See _make_health_result.
-
-    Severity routing:
-      - timeout-sanity (timeout < 60): always `warnings`
-      - wrapper-auth-MISSING / WRONG-MODE: REQUIRED → `issues` (fatal,
-        matches wrapper exit 1); OPTIONAL → demoted via metadata-aware
-        `health_results_ok` at exit-code time. This function reports
-        the auth fault into `issues` unconditionally; demotion happens
-        in health_results_ok based on preset_metadata.
-    """
-    issues = []
-    warnings = []
-    command = None
-    executable = None
-    timeout = None
-    cli = None
-    timeout_warning = False
-
-    try:
-        command = parse_provider_command(provider, provider_config)
-    except ValueError as exc:
-        issues.append(str(exc))
-
-    if command:
-        executable, issue = executable_health(command, root)
-        if issue:
-            issues.append(issue)
-        if executable is not None:
-            cli = shlex.join([executable, *command[1:]])
-
-    if isinstance(provider_config, dict):
-        try:
-            timeout = provider_timeout(provider, provider_config)
-        except ValueError as exc:
-            issues.append(str(exc))
-
-    if timeout is not None and timeout < _MIN_TIMEOUT_WARNING_SECONDS:
-        warnings.append(
-            f"timeout {timeout}s is below {_MIN_TIMEOUT_WARNING_SECONDS}s "
-            f"minimum recommended"
-        )
-        timeout_warning = True
-
-    return _make_health_result(
-        provider,
-        ok=not issues,
-        issues=issues,
-        warnings=warnings,
-        executable=executable,
-        timeout=timeout,
-        cli=cli,
-        auth=None,
-        timeout_warning=timeout_warning,
-    )
-
-
-def provider_health_results(config, providers, root):
-    provider_configs = config.get("providers", {})
-    if not isinstance(provider_configs, dict):
-        return [
-            _make_health_result(
-                provider,
-                ok=False,
-                issues=["providers config must be an object"],
-            )
-            for provider in providers
-        ]
-
-    results = []
-    for provider in providers:
-        if provider not in provider_configs:
-            results.append(
-                _make_health_result(
-                    provider,
-                    ok=False,
-                    issues=["unknown provider"],
-                )
-            )
-            continue
-        results.append(provider_health(provider, provider_configs[provider], root))
-    return results
-
-
-def _format_provider_block(result, *, optional=False):
-    """Render one provider's verbose block. Helper for format_health_results.
-
-    When optional=True, NOT-OK providers render as `WARN` instead of
-    `FAIL` because health_results_ok demotes their issues to warnings
-    (codex bot finding on PR #68 round 3).
-    """
-    provider = result["provider"]
-    fail_label = "WARN" if optional else "FAIL"
-    lines = []
-    if result.get("ok"):
-        first = (
-            f"{provider}: OK - {result.get('executable')} "
-            f"(timeout {result.get('timeout')}s)"
-        )
-    else:
-        # First non-fatal warning OR first issue feeds the headline.
-        if result.get("issues"):
-            first = f"{provider}: {fail_label} - {result['issues'][0]}"
-        else:
-            first = (
-                f"{provider}: OK - {result.get('executable')} "
-                f"(timeout {result.get('timeout')}s)"
-            )
-    lines.append(first)
-    if result.get("cli"):
-        lines.append(f"    cli: {result['cli']}")
-    auth = result.get("auth")
-    if auth is not None:
-        if auth["source"] == "env":
-            lines.append(f"    auth: env ${auth['env_var']} set")
-        elif auth["source"] == "file" and auth["mode_ok"]:
-            lines.append(f"    auth: file {auth['file']} (mode ok)")
-        elif auth["source"] == "file":
-            lines.append(
-                f"    auth: file {auth['file']} (mode WRONG, expected 600/400)"
-            )
-        else:
-            lines.append(
-                f"    auth: missing — ${auth['env_var']} unset, {auth['file']} absent"
-            )
-    probe = result.get("live_probe")
-    if probe is not None:
-        if probe["status"] == "pass":
-            lines.append("    live: pass")
-        else:
-            lines.append(f"    live: FAIL - {probe['cause']}: {probe['detail']}")
-    for w in result.get("warnings", []):
-        lines.append(f"    warning: {w}")
-    for issue in result.get("issues", []):
-        if first.startswith(f"{provider}: {fail_label} - {issue}"):
-            continue  # already in headline
-        lines.append(f"    issue: {issue}")
-    return lines
-
-
-def format_health_results(
-    results, *, env_pollution=None, preset_metadata=None, verbose=False
-):
-    """Render provider health results.
-
-    Default (verbose=False) preserves the existing single-line shape
-    `{provider}: OK - {executable} (timeout Ns)` per provider — used by
-    `cmd_review --check-providers` (codex.py:1397-1403). Verbose mode
-    (cmd_doctor) adds REQUIRED/OPTIONAL split, indented detail rows, and
-    an ENV POLLUTION section. First line per provider stays grep-compatible.
-    """
-    if not verbose:
-        lines = []
-        for result in results:
-            provider = result["provider"]
-            if result["ok"]:
-                lines.append(
-                    f"{provider}: OK - {result['executable']} "
-                    f"(timeout {result['timeout']}s)"
-                )
-                continue
-            for issue in result["issues"]:
-                lines.append(f"{provider}: FAIL - {issue}")
-        return "\n".join(lines)
-
-    # Verbose mode.
-    optional_set = set()
-    if preset_metadata is not None:
-        # Subtract REQUIRED to match health_results_ok demotion logic
-        # (codex bot finding on PR #68): a provider in BOTH lists is
-        # treated as required.
-        required_set = set(preset_metadata.get("providers", []))
-        optional_set = set(preset_metadata.get("optional_providers", [])) - required_set
-
-    required_results = [r for r in results if r["provider"] not in optional_set]
-    optional_results = [r for r in results if r["provider"] in optional_set]
-
-    lines = []
-    if preset_metadata is not None and (required_results or optional_results):
-        if required_results:
-            lines.append("REQUIRED:")
-            for r in required_results:
-                lines.extend(_format_provider_block(r))
-                lines.append("")
-        if optional_results:
-            lines.append("OPTIONAL:")
-            for r in optional_results:
-                lines.extend(_format_provider_block(r, optional=True))
-                lines.append("")
-    else:
-        # No preset → single PROVIDERS section.
-        for r in results:
-            lines.extend(_format_provider_block(r))
-            lines.append("")
-
-    if env_pollution:
-        lines.append(
-            "ENV POLLUTION (would leak into provider spawn pre-TRN-3023; now stripped at spawn):"
-        )
-        for key, redacted in env_pollution:
-            if redacted:
-                lines.append(
-                    f"    ⚠️  {key} set in shell ({redacted}) — stripped at spawn"
-                )
-            else:
-                lines.append(f"    ⚠️  {key} set in shell — stripped at spawn")
-        lines.append("")
-
-    return "\n".join(lines).rstrip("\n")
-
-
-def health_results_ok(results, *, preset_metadata=None):
-    """TRN-3021: metadata-aware exit-code decision.
-
-    REQUIRED providers with `issues` non-empty → False (exit 1).
-    OPTIONAL providers with `issues` non-empty → demoted (don't fail).
-    `preset_metadata=None` (existing call sites in cmd_review --check-providers)
-    treats all providers as REQUIRED, preserving current semantics.
-    """
-    optional_set = set()
-    if preset_metadata is not None:
-        # Subtract REQUIRED first: if a provider appears in BOTH
-        # `providers` (required) and `optional_providers`, the required
-        # designation wins (resolve_preset_providers selects it as
-        # required first). Codex bot finding on PR #68: without this
-        # subtraction, an overlap silently demotes a fatal issue to
-        # warning-only and exits 0.
-        required_set = set(preset_metadata.get("providers", []))
-        optional_set = set(preset_metadata.get("optional_providers", [])) - required_set
-    for result in results:
-        if result.get("issues") and result["provider"] not in optional_set:
-            return False
-    return True
-
-
-def scope_pathspec(root, scope):
-    if not scope or scope == ".":
-        return []
-    candidate = (root / scope).resolve()
-    try:
-        candidate.relative_to(root)
-    except ValueError:
-        return []
-    if candidate.exists():
-        return [scope]
-    return []
-
-
-def git_diff(root, pathspec):
-    args = ["diff", "--no-ext-diff", "--binary", "HEAD", "--"]
-    return run_git(root, args + pathspec, allow_error=True)
-
-
-def git_diff_range(root, base, head, pathspec):
-    args = ["diff", "--no-ext-diff", "--binary", f"{base}...{head}", "--"]
-    return run_git(root, args + pathspec)
-
-
-def changed_paths(root, pathspec):
-    args = ["diff", "--name-only", "--diff-filter=ACMRT", "HEAD", "--"]
-    output = run_git(root, args + pathspec, allow_error=True)
-    return [line for line in output.splitlines() if line.strip()]
-
-
-def changed_paths_range(root, base, head, pathspec):
-    args = ["diff", "--name-only", "--diff-filter=ACMRT", f"{base}...{head}", "--"]
-    output = run_git(root, args + pathspec)
-    return [line for line in output.splitlines() if line.strip()]
-
-
-def untracked_paths(root, pathspec):
-    args = ["ls-files", "--others", "--exclude-standard", "--"]
-    output = run_git(root, args + pathspec, allow_error=True)
-    return [line for line in output.splitlines() if line.strip()]
-
-
-def is_text_file(path):
-    try:
-        sample = path.read_bytes()[:4096]
-    except OSError:
-        return False
-    return b"\0" not in sample
-
-
-def read_text_file(path, max_bytes=200000):
-    try:
-        data = path.read_bytes()
-    except OSError as exc:
-        return f"[unreadable: {exc}]"
-    if b"\0" in data:
-        return "[binary file omitted]"
-    if len(data) > max_bytes:
-        head = data[:max_bytes].decode("utf-8", errors="replace")
-        return head + "\n[truncated: file exceeds max snapshot bytes]\n"
-    return data.decode("utf-8", errors="replace")
-
-
-def read_text_file_at_ref(root, ref, rel, max_bytes=200000):
-    result = subprocess.run(
-        ["git", "show", f"{ref}:{rel}"],
-        cwd=str(root),
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-    )
-    if result.returncode != 0:
-        return None
-    data = result.stdout
-    if b"\0" in data:
-        return "[binary file omitted]"
-    if len(data) > max_bytes:
-        head = data[:max_bytes].decode("utf-8", errors="replace")
-        return head + "\n[truncated: file exceeds max snapshot bytes]\n"
-    return data.decode("utf-8", errors="replace")
-
-
-def synthetic_untracked_diff(root, paths):
-    chunks = []
-    for rel in paths:
-        path = root / rel
-        if not path.is_file() or not is_text_file(path):
-            chunks.append(f"diff --git a/{rel} b/{rel}\nnew file mode 100644\n")
-            chunks.append(
-                "--- /dev/null\n+++ b/{0}\n[binary or unreadable]\n".format(rel)
-            )
-            continue
-        lines = read_text_file(path).splitlines(keepends=True)
-        chunks.extend(
-            difflib.unified_diff(
-                [],
-                lines,
-                fromfile="/dev/null",
-                tofile=f"b/{rel}",
-            )
-        )
-    return "".join(chunks)
-
-
-def file_snapshots(root, paths):
-    if not paths:
-        return "(no changed or untracked file snapshots)\n"
-
-    chunks = []
-    seen = set()
-    for rel in paths:
-        if rel in seen:
-            continue
-        seen.add(rel)
-        path = root / rel
-        chunks.append(f"### {rel}\n\n")
-        if not path.exists():
-            chunks.append("[deleted]\n\n")
-            continue
-        if path.is_dir():
-            chunks.append("[directory omitted]\n\n")
-            continue
-        content = read_text_file(path)
-        chunks.append("```text\n")
-        chunks.append(content)
-        if content and not content.endswith("\n"):
-            chunks.append("\n")
-        chunks.append("```\n\n")
-    return "".join(chunks)
-
-
-def file_snapshots_at_ref(root, paths, ref):
-    if not paths:
-        return "(no changed file snapshots)\n"
-
-    chunks = []
-    seen = set()
-    for rel in paths:
-        if rel in seen:
-            continue
-        seen.add(rel)
-        chunks.append(f"### {rel}\n\n")
-        content = read_text_file_at_ref(root, ref, rel)
-        if content is None:
-            chunks.append("[deleted]\n\n")
-            continue
-        chunks.append("```text\n")
-        chunks.append(content)
-        if content and not content.endswith("\n"):
-            chunks.append("\n")
-        chunks.append("```\n\n")
-    return "".join(chunks)
-
-
-def review_input_sha256(diff, files):
-    payload = json.dumps(
-        {"diff": diff, "files": files},
-        sort_keys=True,
-        separators=(",", ":"),
-        ensure_ascii=False,
-    )
-    return hashlib.sha256(payload.encode("utf-8")).hexdigest()
-
-
-def make_review_input(diff, files, metadata):
-    metadata = dict(metadata)
-    metadata["review_input_sha256"] = review_input_sha256(diff, files)
-    return {"diff": diff, "files": files, "metadata": metadata}
-
-
-def working_tree_review_input(root, scope):
-    pathspec = scope_pathspec(root, scope)
-    tracked_diff = git_diff(root, pathspec)
-    untracked = untracked_paths(root, pathspec)
-    untracked_diff = synthetic_untracked_diff(root, untracked)
-    diff = (
-        tracked_diff
-        + ("\n" if tracked_diff and untracked_diff else "")
-        + (untracked_diff)
-    )
-    paths = changed_paths(root, pathspec) + untracked
-    files = file_snapshots(root, paths)
-    if not diff.strip():
-        diff = "(no tracked or untracked git diff)"
-    return make_review_input(
-        diff,
-        files,
-        {
-            "mode": "working-tree",
-            "base": "HEAD",
-            "head": "working-tree",
-            "pr": None,
-            "scope": scope or ".",
-            "changed_paths": paths,
-            "snapshot_source": "working-tree",
-        },
-    )
-
-
-def base_head_review_input(root, base, head, scope):
-    pathspec = scope_pathspec(root, scope)
-    try:
-        diff = git_diff_range(root, base, head, pathspec)
-        paths = changed_paths_range(root, base, head, pathspec)
-    except RuntimeError as exc:
-        raise SystemExit(
-            f"trinity-codex: unable to collect git diff for {base}...{head}: {exc}"
-        ) from exc
-    files = file_snapshots_at_ref(root, paths, head)
-    if not diff.strip():
-        diff = f"(no git diff for {base}...{head})"
-    return make_review_input(
-        diff,
-        files,
-        {
-            "mode": "base-head",
-            "base": base,
-            "head": head,
-            "pr": None,
-            "scope": scope or ".",
-            "changed_paths": paths,
-            "snapshot_source": f"git:{head}",
-        },
-    )
-
-
-def run_gh(root, args, label):
-    try:
-        result = subprocess.run(
-            ["gh"] + args,
-            cwd=str(root),
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            universal_newlines=True,
-        )
-    except FileNotFoundError:
-        raise SystemExit(
-            "trinity-codex: gh command not found; install/authenticate gh for --pr"
-        )
-    if result.returncode != 0:
-        detail = (result.stderr or result.stdout).strip() or "gh command failed"
-        raise SystemExit(f"trinity-codex: {label} failed: {detail}")
-    return result.stdout
-
-
-def gh_pr_view(root, pr_number):
-    output = run_gh(
-        root,
-        [
-            "pr",
-            "view",
-            str(pr_number),
-            "--json",
-            "number,url,baseRefName,headRefName,headRefOid,files",
-        ],
-        f"gh pr view {pr_number}",
-    )
-    try:
-        return json.loads(output)
-    except json.JSONDecodeError as exc:
-        raise SystemExit(f"trinity-codex: invalid gh pr view JSON: {exc}")
-
-
-def gh_pr_diff(root, pr_number):
-    return run_gh(
-        root,
-        ["pr", "diff", str(pr_number), "--patch", "--color", "never"],
-        f"gh pr diff {pr_number}",
-    )
-
-
-def git_commit_exists(root, ref):
-    result = subprocess.run(
-        ["git", "cat-file", "-e", f"{ref}^{{commit}}"],
-        cwd=str(root),
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-    )
-    return result.returncode == 0
-
-
-def ensure_pr_head_available(root, pr_number, head_oid):
-    if not head_oid:
-        return None
-    if git_commit_exists(root, head_oid):
-        return head_oid
-    run_git(
-        root,
-        ["fetch", "--quiet", "origin", f"pull/{pr_number}/head"],
-        allow_error=True,
-    )
-    if git_commit_exists(root, head_oid):
-        return head_oid
-    return None
-
-
-def pr_changed_paths(view):
-    paths = []
-    for item in view.get("files", []) or []:
-        path = item.get("path")
-        if path:
-            paths.append(path)
-    return paths
-
-
-def pr_review_input(root, pr_number, scope):
-    view = gh_pr_view(root, pr_number)
-    diff = gh_pr_diff(root, pr_number)
-    paths = pr_changed_paths(view)
-    head_oid = view.get("headRefOid")
-    snapshot_ref = ensure_pr_head_available(root, pr_number, head_oid)
-    if snapshot_ref:
-        files = file_snapshots_at_ref(root, paths, snapshot_ref)
-        snapshot_source = f"git:{snapshot_ref}"
-    else:
-        files = "(PR head snapshots unavailable locally)\n"
-        snapshot_source = "unavailable"
-    if not diff.strip():
-        diff = f"(no GitHub PR diff for PR #{pr_number})"
-    return make_review_input(
-        diff,
-        files,
-        {
-            "mode": "pr",
-            "base": view.get("baseRefName"),
-            "head": view.get("headRefName"),
-            "head_oid": head_oid,
-            "pr": int(view.get("number") or pr_number),
-            "pr_url": view.get("url"),
-            "scope": scope or ".",
-            "changed_paths": paths,
-            "snapshot_source": snapshot_source,
-        },
-    )
-
-
-def resolve_review_input(args, root):
-    if args.pr is not None:
-        return pr_review_input(root, args.pr, args.scope)
-    if args.base or args.head:
-        if not args.base or not args.head:
-            raise SystemExit("trinity-codex: --base and --head must be used together")
-        return base_head_review_input(root, args.base, args.head, args.scope)
-    return working_tree_review_input(root, args.scope)
-
-
-def normalize_review_doc_id(value, flag):
-    normalized = (value or "").strip().upper()
-    if not re.fullmatch(r"[A-Z]{3}-\d{4}", normalized):
-        raise SystemExit(f"trinity: {flag} must look like COR-1602")
-    return normalized
-
-
-def strict_review_metadata(sop, rubric, template):
-    return {
-        "enabled": True,
-        "sop": sop,
-        "rubric": rubric,
-        "pass_threshold": template["pass_threshold"],
-        "calibration": template["calibration"],
-        "decision_rule": template["decision_rule"],
-        "output_schema": list(STRICT_REVIEW_OUTPUT_SCHEMA),
-    }
-
-
-def resolve_strict_review(args):
-    if bool(args.sop) != bool(args.rubric):
-        raise SystemExit("trinity: --sop and --rubric must be used together")
-    if not args.sop and not args.rubric:
-        return None
-
-    sop = normalize_review_doc_id(args.sop, "--sop")
-    rubric = normalize_review_doc_id(args.rubric, "--rubric")
-    template = STRICT_REVIEW_TEMPLATES.get((sop, rubric))
-    if template is None:
-        raise SystemExit(
-            f"trinity: unsupported strict review template: SOP {sop} with rubric {rubric}"
-        )
-    metadata = strict_review_metadata(sop, rubric, template)
-    return {**metadata, "template": template}
-
-
-def render_strict_review_instructions(strict_review):
-    template = strict_review["template"]
-    lines = [
-        "## Strict COR Review Mode",
-        "",
-        f"SOP: {strict_review['sop']}",
-        f"Rubric: {strict_review['rubric']} ({template['rubric_title']})",
-        f"Calibration: {strict_review['calibration']}",
-        f"PASS threshold: {strict_review['pass_threshold']:.1f}/10",
-        "",
-        "Follow the SOP workflow and score the artifact with this rubric:",
-        "",
-        "| Criterion | Weight | Scoring focus |",
-        "|-----------|--------|---------------|",
-    ]
-    for criterion, weight, focus in template["criteria"]:
-        lines.append(f"| {criterion} | {weight} | {focus} |")
-
-    matrix_rows = [
-        f"| {criterion} | {weight} | X.X | ... |"
-        for criterion, weight, _focus in template["criteria"]
-    ]
-    lines.extend(
-        [
-            "",
-            template["non_code_note"],
-            "",
-            "Use COR-1611 calibration: cite every deduction, distinguish blocking "
-            "findings from advisory improvements, and reserve 10/10 for zero "
-            "remaining improvements.",
-            "",
-            "Required output:",
-            "",
-            "### Findings",
-            "",
-            "- List each finding with severity, file/path reference when applicable, "
-            "and whether it is blocking or advisory.",
-            "",
-            "### Decision Matrix",
-            "",
-            "| Criterion | Weight | Score | Rationale |",
-            "|-----------|--------|-------|-----------|",
-            *matrix_rows,
-            "",
-            "**Weighted Average: X.X/10 - PASS/FIX**",
-            "",
-            template["decision_rule"],
-        ]
-    )
-    return "\n".join(lines) + "\n"
-
-
-def render_prompt(
-    config, root, scope, review_input, strict_review=None, *, task_type=None
-):
-    diff = review_input["diff"]
-    files = review_input["files"]
-    template = config.get("review", {}).get("prompt_template", "{diff}\n\n{files}\n")
-    prompt = (
-        template.replace("{scope}", scope or ".")
-        .replace("{root}", str(root))
-        .replace("{diff}", diff)
-        .replace("{files}", files)
-    )
-    if strict_review is None:
-        base = REVIEW_ONLY_INSTRUCTION + "\n" + prompt
-    else:
-        base = (
-            render_strict_review_instructions(strict_review)
-            + "\n"
-            + REVIEW_ONLY_INSTRUCTION
-            + "\n## Review Artifact\n\n"
-            + prompt
-        )
-    # TRN-3022: schema addendum appended AFTER all other sections so
-    # the "LAST in your output" instruction is truly at the bottom.
-    return base + _review_schema_addendum(task_type, strict_review=strict_review)
-
-
-def slugify(value):
-    value = value.strip() or "review"
-    value = re.sub(r"[^A-Za-z0-9._-]+", "-", value)
-    return value.strip("-") or "review"
-
-
-def make_review_dir(out_dir, scope):
-    stamp = dt.datetime.now().strftime("%Y%m%d-%H%M%S")
-    base = Path(out_dir).expanduser()
-    slug = slugify(scope)
-    for index in range(100):
-        suffix = f"{stamp}-{slug}" if index == 0 else f"{stamp}-{slug}-{index}"
-        review_dir = base / suffix
-        try:
-            review_dir.mkdir(parents=True, exist_ok=False)
-        except FileExistsError:
-            continue
-        (review_dir / "raw").mkdir()
-        (review_dir / "logs").mkdir()
-        return review_dir
-    raise SystemExit("trinity-codex: unable to create unique review directory")
-
-
-def review_parallelism(config, providers):
-    review_config = review_section(config)
-    raw_value = review_config.get("max_parallel_providers")
-    if raw_value is None:
-        return len(providers)
-    if isinstance(raw_value, bool) or not isinstance(raw_value, int):
-        raise SystemExit(
-            "trinity: review.max_parallel_providers must be an integer "
-            f"between 1 and {len(providers)}"
-        )
-    if raw_value < 1 or raw_value > len(providers):
-        raise SystemExit(
-            f"trinity: review.max_parallel_providers must be between 1 and {len(providers)}"
-        )
-    return raw_value
-
-
-def ordered_subset(providers, selected):
-    selected = set(selected)
-    ordered = [provider for provider in providers if provider in selected]
-    ordered.extend(sorted(selected.difference(providers)))
-    return ordered
-
-
-def ordered_cleanup(providers, cleanup):
-    payload = {
-        provider: cleanup[provider] for provider in providers if provider in cleanup
-    }
-    for provider in sorted(set(cleanup).difference(payload)):
-        payload[provider] = cleanup[provider]
-    return payload
-
-
-def write_incomplete(
-    review_dir,
-    status,
-    providers,
-    started_providers,
-    cleanup,
-    message=None,
-):
-    cleanup_payload = ordered_cleanup(providers, cleanup)
-    payload = {
-        "status": status,
-        "timestamp": timestamp(),
-        "review_dir": str(review_dir),
-        "providers_selected": providers,
-        "providers_started": ordered_subset(providers, started_providers),
-        "providers_running_at_cleanup": ordered_subset(providers, cleanup),
-        "cleanup": cleanup_payload,
-    }
-    if message:
-        payload["message"] = message
-    (review_dir / "incomplete.json").write_text(
-        json.dumps(payload, indent=2, ensure_ascii=False) + "\n"
-    )
-
-
-def _sanitize_md(text):
-    """Strip embedded newlines from Markdown-rendered finding fields."""
-    return text.replace("\n", " ") if "\n" in text else text
-
-
-def _render_finding_items(label, items):
-    """Render a labeled list of finding items (blocking or advisories).
-
-    Returns a list of lines including the **Label:** header, one bullet per
-    item, and a trailing blank line. Returns an empty list if items is empty.
-    """
-    if not items:
-        return []
-    lines = [f"**{label}:**"]
-    for item in items:
-        title = _sanitize_md(item.get("title", ""))
-        evidence = _sanitize_md(item.get("evidence", ""))
-        fix = item.get("fix")
-        if fix:
-            fix = _sanitize_md(fix)
-        if evidence:
-            line = f"- **{title}** at `{evidence}`"
-        else:
-            line = f"- **{title}** (no evidence cited)"
-        if fix:
-            line += f" — {fix}"
-        lines.append(line)
-    lines.append("")
-    return lines
-
-
-def _render_findings_for(result, parsed):
-    """Render a single provider's structured findings block.
-
-    Returns a list of lines for the Findings section, or an empty list
-    if the provider has no findings to render (structured-but-clean).
-    """
-    provider = result["provider"]
-    eff = parsed["effective_decision"]
-    score = parsed["weighted_score"]
-    blocking = parsed.get("blocking", [])
-    advisories = parsed.get("advisories", [])
-    lines = []
-    lines.append(
-        f"### {provider} — {eff} "
-        f"({score}, {len(blocking)} blocking, {len(advisories)} advisories)"
-    )
-    lines.append("")
-    lines.extend(_render_finding_items("Blocking", blocking))
-    lines.extend(_render_finding_items("Advisories", advisories))
-    return lines
-
-
-def _compute_summary(results, parsed_per_provider):
-    """Aggregate counts + verdict from per-provider results.
-
-    Returns dict with: verdict, n_pass, n_fix, n_fail, total, mean_score
-    (or None), n_blocking, n_advisories, convergence_count.
-    """
-    n_pass = n_fix = n_fail = 0
-    scores = []
-    n_blocking = 0
-    n_advisories = 0
-    title_provider_pairs = []  # (title, provider_name) — for convergence
-
-    for i, result in enumerate(results):
-        rc = result.get("returncode", -1)
-        if rc != 0:
-            n_fail += 1
-            continue
-        parsed = parsed_per_provider[i]
-        if parsed is None:
-            n_pass += 1  # legacy rc=0 PASS
-            continue
-        eff = parsed.get("effective_decision", parsed.get("decision"))
-        if eff == "FIX":
-            n_fix += 1
-        else:
-            n_pass += 1
-        scores.append(parsed.get("weighted_score"))
-        prov = result["provider"]
-        for item in parsed.get("blocking", []):
-            n_blocking += 1
-            title = (item.get("title") or "").strip()
-            if title:
-                title_provider_pairs.append((title, prov))
-        for item in parsed.get("advisories", []):
-            n_advisories += 1
-            title = (item.get("title") or "").strip()
-            if title:
-                title_provider_pairs.append((title, prov))
-
-    total = len(results)
-    has_structured = any(p is not None for p in parsed_per_provider)
-
-    # Convergence: distinct titles appearing across >=2 distinct providers.
-    title_to_providers = {}
-    for title, prov in title_provider_pairs:
-        title_to_providers.setdefault(title, set()).add(prov)
-    convergence_count = sum(
-        1 for provs in title_to_providers.values() if len(provs) >= 2
-    )
-
-    # Verdict precedence: INCONCLUSIVE > LEGACY > NEEDS_FIXES > ALL_PASS.
-    if n_fail > 0:
-        verdict = "INCONCLUSIVE"
-    elif not has_structured:
-        verdict = "LEGACY"
-    elif n_fix > 0:
-        verdict = "NEEDS_FIXES"
-    else:
-        verdict = "ALL_PASS"
-
-    valid_scores = [s for s in scores if isinstance(s, (int, float))]
-    mean_score = sum(valid_scores) / len(valid_scores) if valid_scores else None
-
-    return {
-        "verdict": verdict,
-        "n_pass": n_pass,
-        "n_fix": n_fix,
-        "n_fail": n_fail,
-        "total": total,
-        "mean_score": mean_score,
-        "n_blocking": n_blocking,
-        "n_advisories": n_advisories,
-        "convergence_count": convergence_count,
-    }
-
-
-def _render_summary_block(summary):
-    """Render the aggregate Summary section as markdown lines."""
-    lines = ["## Summary", ""]
-    lines.append(f"- **Verdict**: {summary['verdict']}")
-    providers_line = (
-        f"{summary['n_pass']}/{summary['total']} PASS · "
-        f"{summary['n_fix']} FIX · "
-        f"{summary['n_fail']} FAIL"
-    )
-    if summary["mean_score"] is not None:
-        providers_line += f" (mean score {summary['mean_score']:.2f})"
-    lines.append(f"- **Providers**: {providers_line}")
-    if summary["verdict"] == "LEGACY":
-        lines.append("- **Findings**: —")
-        lines.append("- **Convergence**: —")
-    else:
-        lines.append(
-            f"- **Findings**: {summary['n_blocking']} blocking · "
-            f"{summary['n_advisories']} advisories"
-        )
-        conv = summary["convergence_count"]
-        conv_text = f"{conv} titles flagged by ≥2 providers" if conv else "none"
-        lines.append(f"- **Convergence**: {conv_text}")
-    return lines
-
-
-def _format_cli_summary(summary, synthesis_path):
-    """One-line stderr summary for cmd_review completion."""
-    providers = (
-        f"{summary['n_pass']}/{summary['total']} PASS · "
-        f"{summary['n_fix']} FIX · "
-        f"{summary['n_fail']} FAIL"
-    )
-    if summary["mean_score"] is not None:
-        providers += f" (mean {summary['mean_score']:.2f})"
-    return (
-        f"trinity review: {summary['verdict']} — "
-        f"{providers} — synthesis: {synthesis_path}"
-    )
-
-
-def write_synthesis(review_dir, scope, results, strict_review=None):
-    # TRN-3022: parse structured review schema per provider.
-    # rc != 0 → FAIL <rc> regardless of structured content.
-    # rc == 0 + parsed → enriched status.
-    # rc == 0 + no parsed → legacy PASS.
-    parsed_per_provider = []
-    for result in results:
-        rc = result.get("returncode", -1)
-        if rc == 0:
-            raw_path = review_dir / result["raw"]
-            raw_text = _safe_read_raw(raw_path)
-            if raw_text is not None:
-                parsed_per_provider.append(
-                    parse_structured_review(
-                        raw_text,
-                        pass_threshold=strict_review["pass_threshold"]
-                        if strict_review
-                        else None,
-                    )
-                )
-            else:
-                parsed_per_provider.append(None)
-        else:
-            parsed_per_provider.append(None)
-
-    any_structured = any(p is not None for p in parsed_per_provider)
-
-    summary = _compute_summary(results, parsed_per_provider)
-    summary_lines = _render_summary_block(summary)
-
-    # Legacy path: same Provider Status table + Notes shape as pre-TRN-3022,
-    # but TRN-3028 prepends a Summary block (verdict=LEGACY) at the top.
-    if not any_structured:
-        lines = [
-            "# Trinity Review Synthesis",
-            "",
-            f"Scope: {scope or '.'}",
-            "",
-        ]
-        lines.extend(summary_lines)
-        lines.extend(
-            [
-                "",
-                "## Provider Status",
-                "",
-                "| Provider | Status | Raw Output |",
-                "|----------|--------|------------|",
-            ]
-        )
-        for result in results:
-            status = (
-                "PASS" if result["returncode"] == 0 else f"FAIL {result['returncode']}"
-            )
-            lines.append(f"| {result['provider']} | {status} | `{result['raw']}` |")
-        lines.extend(
-            [
-                "",
-                "## Notes",
-                "",
-                "This synthesis is deterministic. Inspect raw provider outputs for findings and conflicts.",
-                "",
-            ]
-        )
-        path = review_dir / "synthesis.md"
-        path.write_text("\n".join(lines) + "\n")
-        return summary, path
-
-    # Enriched path: at least one provider has structured output.
-    lines = [
-        "# Trinity Review Synthesis",
-        "",
-        f"Scope: {scope or '.'}",
-        "",
-    ]
-    lines.extend(summary_lines)
-    lines.extend(
-        [
-            "",
-            "## Provider Status",
-            "",
-            "| Provider | Status | Raw Output |",
-            "|----------|--------|------------|",
-        ]
-    )
-    for i, result in enumerate(results):
-        rc = result.get("returncode", -1)
-        if rc != 0:
-            # Returncode precedence: rc wins regardless of structured content.
-            status = f"FAIL {rc}"
-        else:
-            parsed = parsed_per_provider[i]
-            if parsed is not None:
-                eff = parsed["effective_decision"]
-                score = parsed["weighted_score"]
-                n_blocking = len(parsed.get("blocking", []))
-                if eff == "FIX":
-                    status = f"FIX ({score}, {n_blocking} blocking)"
-                else:
-                    status = f"PASS ({score})"
-            else:
-                status = "PASS"
-        lines.append(f"| {result['provider']} | {status} | `{result['raw']}` |")
-
-    # Findings section — only rc=0 providers with parsed results.
-    has_findings = False
-    for i, result in enumerate(results):
-        parsed = parsed_per_provider[i]
-        rc = result.get("returncode", -1)
-        if rc == 0 and parsed is not None:
-            if not has_findings:
-                lines.append("")
-                lines.append("## Findings")
-                lines.append("")
-                has_findings = True
-            lines.extend(_render_findings_for(result, parsed))
-
-    lines.extend(
-        [
-            "",
-            "## Notes",
-            "",
-            "This synthesis is deterministic. Inspect raw provider outputs for findings and conflicts.",
-            "",
-        ]
-    )
-    path = review_dir / "synthesis.md"
-    path.write_text("\n".join(lines) + "\n")
-    return summary, path
-
-
-def cmd_review(args):
-    progress("preparing review prompt")
-    root = resolve_root(args.root)
-    config = load_config(args.config)
-    providers, preset_metadata, resolver_warnings = resolve_review_providers(
-        args, config
-    )
-    strict_review = resolve_strict_review(args)
-    for warning in resolver_warnings:
-        print(warning, file=sys.stderr)
-    provider_configs = config.get("providers", {})
-    health = provider_health_results(config, providers, root)
-    if args.check_providers:
-        print(format_health_results(health))
-        return 0 if health_results_ok(health) else 1
-    if not health_results_ok(health):
-        print(format_health_results(health), file=sys.stderr)
-        return 1
-    max_workers = review_parallelism(config, providers)
-    review_input = resolve_review_input(args, root)
-
-    out_base = args.out_dir or config.get("review", {}).get(
-        "output_dir", ".trinity/reviews"
-    )
-    out_base = Path(out_base)
-    if not out_base.is_absolute():
-        out_base = root / out_base
-    review_dir = make_review_dir(out_base, args.scope)
-
-    try:
-        prompt = render_prompt(
-            config,
-            root,
-            args.scope,
-            review_input,
-            strict_review,
-            task_type=preset_metadata.get("task_type") if preset_metadata else None,
-        )
-        prompt_path = review_dir / "prompt.md"
-        progress("writing prompt")
-        prompt_path.write_text(prompt)
-        # TRN-2018 M1: write metadata.json with all providers queued BEFORE
-        # run_providers. This gives `trinity status --latest` a live view
-        # while providers run, instead of waiting until completion.
-        # R5 fix (codex R5 P2): strict_review threaded into the initial
-        # atomic write — finalize_metadata mutates only finished_at /
-        # results / status / provider_states, so strict_review survives
-        # through the post-run update without a separate write_text that
-        # would race readers.
-        strict_review_block = (
-            {key: value for key, value in strict_review.items() if key != "template"}
-            if strict_review is not None
-            else None
-        )
-        _rm.init_metadata(
-            review_dir,
-            review_id=review_dir.name,
-            review_dir_str=str(review_dir),
-            providers=providers,
-            preset=preset_metadata,
-            scope=args.scope,
-            root=str(root),
-            input=review_input["metadata"],
-            strict_review=strict_review_block,
-        )
-        results = run_providers(
-            max_workers,
-            providers,
-            provider_configs,
-            prompt_path,
-            review_dir,
-            root,
-        )
-        # TRN-2018 R7 fix (codex R6 P2): per CHG L88-89, top-level
-        # `finished_at` stays null until synthesis.md is written.
-        # finalize_metadata must run AFTER write_synthesis succeeds so
-        # a concurrent `trinity status` poll never observes
-        # status=finished while synthesis.md is missing. If
-        # write_synthesis raises, finalize is skipped; the exception
-        # handler writes incomplete.json and metadata.json stays in
-        # `status=running` for the partial review.
-        progress("writing synthesis")
-        summary, synthesis_path = write_synthesis(
-            review_dir, args.scope, results, strict_review=strict_review
-        )
-        progress("writing metadata")
-        _rm.finalize_metadata(review_dir, results)
-        # TRN-3028: emit the completion line directly to stderr without the
-        # `trinity: ` progress prefix so callers can key off the documented
-        # "trinity review: <verdict> — ..." prefix at the START of the line.
-        print(_format_cli_summary(summary, synthesis_path), file=sys.stderr)
-    except KeyboardInterrupt:
-        started_providers = providers if "results" in locals() else []
-        write_incomplete(review_dir, "interrupted", providers, started_providers, {})
-        print(review_dir)
-        return 130
-    except ReviewInterrupted as exc:
-        write_incomplete(
-            review_dir,
-            "interrupted",
-            providers,
-            exc.started_providers,
-            exc.cleanup,
-        )
-        print(review_dir)
-        return 130
-    except ReviewOrchestrationError as exc:
-        write_incomplete(
-            review_dir,
-            "failed",
-            providers,
-            exc.started_providers,
-            exc.cleanup,
-            str(exc),
-        )
-        print(review_dir)
-        return 1
-    except Exception as exc:
-        started_providers = providers if "results" in locals() else []
-        write_incomplete(
-            review_dir, "failed", providers, started_providers, {}, str(exc)
-        )
-        print(review_dir)
-        return 1
-
-    print(review_dir)
-    return 0 if all(item["returncode"] == 0 for item in results) else 1
-
-
-_LIVE_PROBE_TIMEOUT = 10  # hard 10s timeout for live probes.
-
-
-def _probe_provider(provider, provider_config, root):
-    """Run a minimal live probe against a provider CLI.
-
-    Returns a dict with 'status' ('pass'|'fail') and if fail, 'cause'
-    ('auth'|'quota'|'timeout'|'error') and 'detail'. Returns None when
-    the provider cannot be probed (static check would already have
-    caught it).
-    """
-    try:
-        command = parse_provider_command(provider, provider_config)
-    except ValueError:
-        return None
-    executable, issue = executable_health(command, root)
-    if issue or executable is None:
-        return None
-    env = build_provider_env()
-    try:
-        # Mirror run_provider semantics: the prompt is appended as the final
-        # CLI argument (gemini -p / droid exec / codex exec all take it that
-        # way), the probe runs from the resolved root so relative
-        # executable paths match the static executable_health check, and the
-        # probe gets its own session so a timeout can kill the whole process
-        # group — descendants holding the captured pipes would otherwise
-        # stall communicate() well past the advertised timeout.
-        proc = subprocess.Popen(
-            command + ["Reply with exactly one word: OK"],
-            cwd=str(root),
-            stdin=subprocess.DEVNULL,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            text=True,
-            start_new_session=True,
-            env=env,
-        )
-        try:
-            probe_stdout, probe_stderr = proc.communicate(timeout=_LIVE_PROBE_TIMEOUT)
-        except subprocess.TimeoutExpired:
-            # terminate_process_group no-ops once the leader has exited,
-            # but children that inherited the pipes can keep the group
-            # alive (and communicate() blocked). start_new_session=True
-            # makes the group id equal to proc.pid, so signal the group
-            # directly — covering both the live-leader and exited-leader
-            # cases. The probe is a throwaway call; no graceful TERM
-            # needed.
-            try:
-                os.killpg(proc.pid, signal.SIGKILL)
-            except (ProcessLookupError, PermissionError):
-                pass
-            try:
-                proc.communicate(timeout=1)
-            except (subprocess.TimeoutExpired, OSError, ValueError):
-                pass
-            return {
-                "status": "fail",
-                "cause": "timeout",
-                "detail": f"no response within {_LIVE_PROBE_TIMEOUT}s",
-            }
-        if proc.returncode == 0:
-            return {"status": "pass"}
-
-        combined = ((probe_stdout or "") + "\n" + (probe_stderr or "")).lower()
-        if any(
-            p in combined
-            for p in (
-                "401",
-                "403",
-                "unauthorized",
-                "auth",
-                "api key",
-                "invalid key",
-                "forbidden",
-                "permission",
-            )
-        ):
-            cause = "auth"
-        elif any(
-            p in combined
-            for p in (
-                "429",
-                "quota",
-                "rate limit",
-                "rate_limit",
-                "too many",
-                "insufficient",
-            )
-        ):
-            cause = "quota"
-        else:
-            cause = "error"
-
-        detail = (probe_stdout or "").strip()[:200] or (probe_stderr or "").strip()[
-            :200
-        ]
-        if detail:
-            detail = f"exit code {proc.returncode}: {detail}"
-        else:
-            detail = f"exit code {proc.returncode}"
-        return {"status": "fail", "cause": cause, "detail": detail}
-
-    except OSError as exc:
-        # executable_health already vouched for the file, so a launch-time
-        # OSError (FileNotFoundError from a missing shebang interpreter,
-        # ENOEXEC, etc.) means reviews cannot launch this provider —
-        # surface it as a live failure rather than silently omitting the
-        # probe result for a statically healthy provider.
-        return {
-            "status": "fail",
-            "cause": "error",
-            "detail": f"failed to launch: {exc}",
-        }
-
-
-def cmd_doctor(args):
-    root = resolve_health_root(args.root)
-    config = load_config(args.config)
-    providers, preset_metadata, resolver_warnings = resolve_review_providers(
-        args, config
-    )
-    for warning in resolver_warnings:
-        print(warning, file=sys.stderr)
-    health = provider_health_results(config, providers, root)
-    # TRN-3021: attach wrapper-auth check to results, but only for providers
-    # whose resolved executable matches the canonical wrapper path. Test
-    # fixtures with fake bin scripts won't trigger this, so cmd_review's
-    # preflight (which uses provider_health directly) is unaffected.
-    for h in health:
-        if not _is_canonical_wrapper(h.get("executable"), h["provider"]):
-            continue
-        auth = wrapper_auth_check(h["provider"])
-        if auth is None:
-            continue
-        h["auth"] = auth
-        if auth["source"] == "missing":
-            h["issues"].append(
-                f"auth missing: ${auth['env_var']} unset and {auth['file']} absent"
-            )
-            h["ok"] = False
-        elif auth["source"] == "file" and auth["mode_ok"] is False:
-            h["issues"].append(
-                f"auth file {auth['file']} has wrong mode (expected 600 or 400)"
-            )
-            h["ok"] = False
-    if args.live:
-        provider_configs = config.get("providers", {})
-        if not isinstance(provider_configs, dict):
-            provider_configs = {}
-        for h in health:
-            if not h["ok"]:
-                continue
-            probe = _probe_provider(
-                h["provider"], provider_configs.get(h["provider"], {}), root
-            )
-            h["live_probe"] = probe
-            if probe and probe["status"] == "fail":
-                h["issues"].append(f"live probe: {probe['cause']} — {probe['detail']}")
-                h["ok"] = False
-    env_pollution = detect_env_pollution()
-    print(
-        format_health_results(
-            health,
-            env_pollution=env_pollution,
-            preset_metadata=preset_metadata,
-            verbose=True,
-        )
-    )
-    return 0 if health_results_ok(health, preset_metadata=preset_metadata) else 1
-
-
-# ---------------------------------------------------------------------------
-# `trinity status` — read-only summary of the most recent review
-# (TRN-2028 / GitHub issue #35).
-# ---------------------------------------------------------------------------
-
-
-def _format_elapsed(started_iso, finished_iso):
-    """Compute elapsed seconds between two ISO timestamps, clamped to >= 0.
-
-    `make_review_dir` writes timestamps via `dt.datetime.now().isoformat(...)`
-    (naive local time). On fall-back DST a sub-hour run straddling the
-    transition can produce finished < started → negative elapsed. Clamp to
-    `max(0, ...)` to defend against that without introducing tz handling.
-    Returns None if either timestamp is missing or unparseable.
-    """
-    if not started_iso or not finished_iso:
-        return None
-    try:
-        s = dt.datetime.fromisoformat(started_iso)
-        f = dt.datetime.fromisoformat(finished_iso)
-        # Subtraction inside the try too — mixed naive/offset-aware datetimes
-        # raise TypeError on `f - s`. Current writer (`timestamp()`) emits
-        # naive only, so this is defensive against future schema drift.
-        delta = (f - s).total_seconds()
-    except (TypeError, ValueError):
-        return None
-    return max(0, int(delta))
-
-
-def _format_duration(seconds):
-    """Render seconds as 'Xm YYs' / 'Ys' / '?' for None."""
-    if seconds is None:
-        return "?"
-    if seconds < 60:
-        return f"{seconds}s"
-    return f"{seconds // 60}m {seconds % 60:02d}s"
-
-
-def _format_ago(now_iso, then_iso):
-    """Render a coarse 'N ago' string. Returns '?' if either is missing."""
-    secs = _format_elapsed(then_iso, now_iso)
-    if secs is None:
-        return "?"
-    if secs < 60:
-        return f"{secs}s ago"
-    if secs < 3600:
-        return f"{secs // 60}m ago"
-    if secs < 86400:
-        return f"{secs // 3600}h ago"
-    return f"{secs // 86400}d ago"
-
-
-def _print_incomplete_only_summary(review_dir, incomplete_path):
-    """Render a summary for an interrupted review (no metadata.json yet).
-
-    cmd_review writes incomplete.json from its KeyboardInterrupt /
-    ReviewInterrupted / ReviewOrchestrationError handlers BEFORE the
-    metadata.json write would have happened, so an interrupted review has
-    only incomplete.json. This path renders what's available.
-    """
-    try:
-        incomplete = json.loads(incomplete_path.read_text())
-    except json.JSONDecodeError as exc:
-        print(
-            f"trinity: malformed incomplete.json at {incomplete_path}: {exc}",
-            file=sys.stderr,
-        )
-        return 1
-
-    status = incomplete.get("status", "interrupted")
-    when = incomplete.get("timestamp", "?")
-    providers_sel = incomplete.get("providers_selected", []) or []
-    providers_started = incomplete.get("providers_started", []) or []
-    providers_running = incomplete.get("providers_running_at_cleanup", []) or []
-    cleanup = incomplete.get("cleanup", {}) or {}
-    message = incomplete.get("message")
-
-    # cmd_review writes status="interrupted" (KeyboardInterrupt /
-    # ReviewInterrupted) or status="failed" (ReviewOrchestrationError /
-    # bare Exception). Use the stored status as the top-level label so
-    # orchestration failures aren't misrendered as user interruptions.
-    print(f"Latest review: {review_dir}  ({status} at {when})")
-    print(f"  Status: {status}")
-    print()
-    print(f"  Providers selected: {', '.join(providers_sel) or '(none)'}")
-    print(f"  Providers started:  {', '.join(providers_started) or '(none)'}")
-    if providers_running:
-        print(f"  Running at cleanup: {', '.join(providers_running)}")
-    if cleanup:
-        # cleanup_active_processes() (scripts/codex.py:1050) writes
-        # {"pid": ..., "result": "terminated"|"killed"|"kill_timeout"}
-        # per provider. Read 'result' not 'status'.
-        cleanup_lines = ", ".join(
-            f"{name}: {info if isinstance(info, str) else info.get('result', '?')}"
-            for name, info in cleanup.items()
-        )
-        print(f"  Cleanup: {cleanup_lines}")
-    if message:
-        print(f"  Message: {message}")
-    print()
-    print("  (metadata.json not present — review never reached completion)")
-    return 0
-
-
-def _print_review_summary(review_dir):
-    """Render a one-screen summary of one review directory. Returns rc."""
-    metadata_path = review_dir / "metadata.json"
-    incomplete_path = review_dir / "incomplete.json"
-
-    # An interrupted review never reaches the metadata-write step in
-    # cmd_review — incomplete.json exists by itself with the structural
-    # state. If that's the case, render from incomplete.json alone rather
-    # than bailing with "no metadata", since this is the exact artifact
-    # the user most wants summarized.
-    if not metadata_path.exists() and incomplete_path.exists():
-        return _print_incomplete_only_summary(review_dir, incomplete_path)
-
-    if not metadata_path.exists():
-        print(
-            f"trinity: review in progress or no metadata at {metadata_path}",
-            file=sys.stderr,
-        )
-        return 1
-    try:
-        metadata = json.loads(metadata_path.read_text())
-    except json.JSONDecodeError as exc:
-        print(
-            f"trinity: malformed metadata at {metadata_path}: {exc}",
-            file=sys.stderr,
-        )
-        return 1
-
-    # All field reads are defensive — pre-TRN-2019 metadata may lack newer
-    # keys (preset, skipped_optional_providers, input.mode).
-    scope = metadata.get("scope", "?")
-    mode = metadata.get("input", {}).get("mode", "?")
-    preset = metadata.get("preset", {}).get("resolved", "?")
-    skipped = metadata.get("preset", {}).get("skipped_optional_providers", [])
-    results = metadata.get("results", [])
-
-    # "started Xm ago" — earliest started_at across results, vs now.
-    # TRN-2018 R3 fix (codex R2 P2): when M1 metadata is mid-review (results
-    # still empty because run_providers hasn't finished), fall back to the
-    # top-level `started_at` written by init_metadata so the header doesn't
-    # render "started ?" in the exact state the live-status feature is for.
-    started_isos = [r.get("started_at") for r in results if r.get("started_at")]
-    earliest = min(started_isos) if started_isos else metadata.get("started_at")
-    now_iso = dt.datetime.now().isoformat(timespec="seconds")
-    ago = _format_ago(now_iso, earliest) if earliest else "?"
-
-    print(f"Latest review: {review_dir}  (started {ago})")
-    print(f"  Scope: {scope}   Mode: {mode}   Preset: {preset}")
-    print()
-    # TRN-2018 M1: live provider_states section. Rendered when the metadata
-    # was written by init_metadata/update_provider_state (M1+ reviews).
-    # Pre-M1 metadata lacks this key; rendering falls through to the legacy
-    # `Providers:` section below.
-    provider_states = metadata.get("provider_states")
-    if provider_states:
-        print("  Live state:")
-        ls_name_width = max([10] + [len(p) for p in provider_states.keys()])
-        for prov, state in provider_states.items():
-            status_str = state.get("status", "?")
-            bits = [f"{prov:{ls_name_width}s}", status_str]
-            pid = state.get("pid")
-            if pid is not None:
-                bits.append(f"pid={pid}")
-            rc = state.get("returncode")
-            if rc is not None:
-                bits.append(f"rc={rc}")
-            # TRN-2018 R4 fix (codex R3 P2): for currently-running providers,
-            # finished_at is absent; use `now` as the end time so the row
-            # shows elapsed-so-far. Terminal states use their finished_at.
-            end_iso = state.get("finished_at") or (
-                now_iso if status_str == "running" else None
-            )
-            ls_elapsed = _format_duration(
-                _format_elapsed(state.get("started_at"), end_iso)
-            )
-            if ls_elapsed and ls_elapsed != "?":
-                bits.append(f"elapsed {ls_elapsed}")
-            print("    " + "  ".join(bits))
-        print()
-    print("  Providers:")
-    if not results:
-        print("    (no results in metadata)")
-    # Column width = max provider name length (or 10, whichever is larger).
-    # Defends against `claude-code` (11 chars) and any future longer name
-    # shifting the rest of the columns out of alignment.
-    name_width = max([10] + [len(r.get("provider", "?")) for r in results])
-    for r in results:
-        name = r.get("provider", "?")
-        rc = r.get("returncode")
-        rc_str = "?" if rc is None else f"rc={rc}"
-        marker = "✓" if rc == 0 else "✗"
-        elapsed = _format_elapsed(r.get("started_at"), r.get("finished_at"))
-        elapsed_str = _format_duration(elapsed)
-        suffix = " (timeout)" if rc == 124 else ""
-        print(
-            f"    {name:{name_width}s} {marker} {rc_str:8s} elapsed {elapsed_str}{suffix}"
-        )
-    print()
-    if skipped:
-        skipped_str = ", ".join(
-            f"{s.get('provider', '?')} ({s.get('reason', '?')})" for s in skipped
-        )
-        print(f"  Skipped optional: {skipped_str}")
-    synthesis_path = review_dir / "synthesis.md"
-    incomplete_path = review_dir / "incomplete.json"
-    if synthesis_path.exists():
-        print("  Synthesis: ✓ synthesis.md")
-    else:
-        print("  Synthesis: missing")
-    incomplete_status = None
-    if incomplete_path.exists():
-        try:
-            incomplete = json.loads(incomplete_path.read_text())
-            incomplete_status = incomplete.get("status", "interrupted")
-        except (json.JSONDecodeError, OSError):
-            incomplete_status = "interrupted"
-        print(f"  Incomplete: ✓ incomplete.json (status={incomplete_status})")
-
-    # Final status line.
-    if incomplete_status is not None:
-        # Use the stored status directly. cmd_review writes status=
-        # "interrupted" (KeyboardInterrupt / ReviewInterrupted) or
-        # "failed" (ReviewOrchestrationError / bare Exception) — see
-        # scripts/codex.py:1370-1402. Hard-coding "interrupted" here
-        # would mislabel an exception-after-metadata-write (e.g. inside
-        # write_synthesis()) as a user interruption.
-        overall = incomplete_status
-    elif metadata.get("status") == "running":
-        # TRN-2018 R4 fix (codex R3 P2): when M1's top-level status is
-        # `running`, a partial `results[]` (one provider finished while
-        # another is still queued/running) must NOT report `completed`.
-        # The live state is authoritative — defer to it until terminal.
-        overall = "running"
-    elif metadata.get("status") in ("failed", "timed_out", "finished"):
-        # M1 terminal: top-level status is authoritative (correctly accounts
-        # for failed/timed_out providers via finalize_metadata precedence).
-        overall = metadata["status"]
-    elif not results:
-        # Pre-M1 with no results (or M1 without a top-level status set,
-        # which shouldn't happen but is defended).
-        overall = "unknown (no results)"
-    elif any(r.get("returncode") != 0 for r in results):
-        # Pre-M1 path: any non-success rc → partial.
-        overall = "partial"
-    else:
-        # Pre-M1 path: all-zero results → completed.
-        overall = "completed"
-    print(f"  Status: {overall}")
-    return 0
-
-
-def cmd_status(args):
-    # Resolve to git top-level (with non-git fallback) so `trinity status`
-    # works from any subdirectory, matching cmd_doctor's behavior. Without
-    # this, a user running from `scripts/` would look at
-    # `scripts/.trinity/reviews/` instead of the repo-root one where
-    # `cmd_review` actually writes artifacts.
-    root = resolve_health_root(args.root)
-    reviews_dir = root / ".trinity" / "reviews"
-    # TRN-2018 M1 behavior change (per CHG L154): missing or empty reviews
-    # dir now exits 0 with `no reviews found` on stdout. Previous behavior:
-    # exit 1 with `no reviews dir at <path>` or `no reviews under <path>`
-    # on stderr. Flagged in CHANGELOG.
-    if not reviews_dir.is_dir():
-        print("no reviews found")
-        return 0
-
-    # Sort key: (timestamp_prefix, mtime). The fixed-width
-    # %Y%m%d-%H%M%S prefix gives chronological order across distinct
-    # seconds; mtime breaks ties for same-second creates where the
-    # slug or numeric collision suffix (`-10` vs `-2`) doesn't preserve
-    # creation order. `make_review_dir()` (scripts/codex.py:914) only
-    # stamps to seconds and appends `<slug>[-<index>]`, so two reviews
-    # made in the same second can have lex order != creation order.
-    def _sort_key(d):
-        try:
-            mtime = d.stat().st_mtime
-        except OSError:
-            mtime = 0.0
-        return (d.name[:15], mtime)
-
-    candidates = sorted(
-        [d for d in reviews_dir.iterdir() if d.is_dir()],
-        key=_sort_key,
-        reverse=True,
-    )
-    if not candidates:
-        print("no reviews found")
-        return 0
-    return _print_review_summary(candidates[0])
 
 
 def cmd_session_path(args):
